version 0.0.5

+++ src/fim.h
@@ -41,7 +41,7 @@ void fim_invert(float * restrict A, const size_t N);
 
 void fim_set_min_to_zero(float * I, size_t N);
 
-int fim_maxAtOrigo(const float * restrict V, const int M, const int N, const int P);
+int fim_maxAtOrigo(const float * restrict V, const int64_t M, const int64_t N, const int64_t P);
   /* Check that the MAX of the fim is in the middle
    * returns 1 on success.
    * Returns 0 if any of the image dimensions are even
@@ -50,35 +50,35 @@ int fim_maxAtOrigo(const float * restrict V, const int M, const int N, const int
 void fim_stats(const float * A, size_t N);
 // Print some info about A to stdout
 
-void fim_flipall(float * restrict T, const float * restrict A, const int a1, const int a2, const int a3);
+void fim_flipall(float * restrict T, const float * restrict A, const int64_t a1, const int64_t a2, const int64_t a3);
   /* 
    * MATLAB:
    * T = flip(flip(flip(A,1),2),3)*/
 
 
-void fim_insert(float * restrict T, const int t1, const int t2, const int t3, 
-    const float * restrict F, const int f1, const int f2, const int f3);
+void fim_insert(float * restrict T, const int64_t t1, const int64_t t2, const int64_t t3, 
+    const float * restrict F, const int64_t f1, const int64_t f2, const int64_t f3);
   /* Insert F [f1xf2xf3] into T [t1xt2xt3] in the "upper left" corner 
    * MATLAB:
    * T(1:size(F,1), 1:size(F,2), 1:sizes(F,3) = F;
    * */
-void fim_insert_ref(float * T, int t1, int t2, int t3, 
-    float * F, int f1, int f2, int f3);
+void fim_insert_ref(float * T, int64_t t1, int64_t t2, int64_t t3, 
+    float * F, int64_t f1, int64_t f2, int64_t f3);
 
-float * fim_get_cuboid(float * restrict A, const int M, const int N, const int P,
-    const int m0, const int m1, const int n0, const int n1, const int p0, const int p1);
+float * fim_get_cuboid(float * restrict A, const int64_t M, const int64_t N, const int64_t P,
+    const int64_t m0, const int64_t m1, const int64_t n0, const int64_t n1, const int64_t p0, const int64_t p1);
 /* MATLAB:
  * Y = A(m0:m1, n0:n1, p0:p1)
  */
 
-float * fim_subregion(float * restrict A, const int M, const int N, const int P, const int m, const int n, const int p);
+float * fim_subregion(float * restrict A, const int64_t M, const int64_t N, const int64_t P, const int64_t m, const int64_t n, const int64_t p);
 /* MATLAB:
  * Y = A(1:m, 1:n, 1:p);
  */
 
-float * fim_subregion_ref(float * A, int M, int N, int P, int m, int n, int p);
+float * fim_subregion_ref(float * A, int64_t M, int64_t N, int64_t P, int64_t m, int64_t n, int64_t p);
 
-void fim_normalize_sum1(float * psf, int M, int N, int P);
+void fim_normalize_sum1(float * psf, int64_t M, int64_t N, int64_t P);
 
 
 float * fim_copy(const float * restrict V, const size_t N);
@@ -91,13 +91,13 @@ float * fim_constant(const size_t N, const float value);
   // Allocate and return an array of N floats sets to a constant value
 
 void fim_circshift(float * restrict A, 
-    const int M, const int N, const int P, 
-    const int sm, const int sn, const int sp);
+    const int64_t M, const int64_t N, const int64_t P, 
+    const int64_t sm, const int64_t sn, const int64_t sp);
   /* Shift the image A [MxNxP] by sm, sn, sp in each dimension */
 
 float * fim_expand(const float * restrict in, 
-    const int pM, const int pN, const int pP, 
-    const int M, const int N, const int P);
+    const int64_t pM, const int64_t pN, const int64_t pP, 
+    const int64_t M, const int64_t N, const int64_t P);
   /* "expand an image" by making it larger 
    * pM, ... current size
    * M, Nm ... new size
@@ -108,15 +108,15 @@ float fim_mse(float * A, float * B, size_t N);
    */
 
 void shift_vector(float * restrict V, 
-    const int S, 
-    const int N,
-    const int k);
+    const int64_t S, 
+    const int64_t N,
+    const int64_t k);
   /* Circular shift of a vector of length N with stride S by step k */
 
 void shift_vector_buf(float * restrict V, 
-    const int S, 
-    const int N,
-    int k, float * restrict buffer);
+    const int64_t S, 
+    const int64_t N,
+    int64_t k, float * restrict buffer);
 
 void fim_mult_scalar(float * I, size_t N, float x);
